+++ src/components/liquidity/LiquidityContainer.js
@@ -10,12 +10,12 @@ import { PactContext } from '../../contexts/PactContext';
 import { ApplicationContext } from '../../contexts/ApplicationContext';
 import { GameEditionContext } from '../../contexts/GameEditionContext';
 import { reduceBalance, getCorrectBalance } from '../../utils/reduceBalance';
-import WalletRequestView from '../../components/modals/swap-modals/WalletRequestView';
+import WalletRequestView from '../../components/modals/WalletRequestView';
 import { ArrowBack, CogIcon } from '../../assets';
 import Label from '../../components/shared/Label';
 import CustomButton from '../../components/shared/CustomButton';
 import ReviewTxModal from '../../components/modals/liquidity/ReviewTxModal';
-import TxView from '../../components/modals/swap-modals/TxView';
+import TxView from '../../components/modals/TxView';
 import tokenData from '../../constants/cryptoCurrencies';
 import SwapForm from '../../components/swap/SwapForm';
 import TokenSelectorModalContent from '../../components/modals/swap-modals/TokenSelectorModalContent';
