+++ src/components/liquidity/RemoveLiqContainer.js
@@ -6,8 +6,8 @@ import { ModalContext } from '../../contexts/ModalContext';
 import { LiquidityContext } from '../../contexts/LiquidityContext';
 import { GameEditionContext } from '../../contexts/GameEditionContext';
 import { WalletContext } from '../../contexts/WalletContext';
-import TxView from '../../components/modals/swap-modals/TxView';
-import WalletRequestView from '../../components/modals/swap-modals/WalletRequestView';
+import TxView from '../../components/modals/TxView';
+import WalletRequestView from '../../components/modals/WalletRequestView';
 import CustomButton from '../../components/shared/CustomButton';
 import FormContainer from '../../components/shared/FormContainer';
 import Input from '../../components/shared/Input';
