+++ src/components/modals/PopupTxView.js
@@ -1,12 +1,12 @@
 import React, { useContext } from 'react';
 import { Icon } from 'semantic-ui-react';
-import { AccountContext } from '../../../contexts/AccountContext';
-import { GameEditionContext } from '../../../contexts/GameEditionContext';
-import { ApplicationContext } from '../../../contexts/ApplicationContext';
-import CopyPopup from '../../../components/shared/CopyPopup';
-import CustomPopup from '../../../components/shared/CustomPopup';
-import { theme } from '../../../styles/theme';
-import Label from '../../shared/Label';
+import { AccountContext } from '../../contexts/AccountContext';
+import { GameEditionContext } from '../../contexts/GameEditionContext';
+import { ApplicationContext } from '../../contexts/ApplicationContext';
+import CopyPopup from '../shared/CopyPopup';
+import CustomPopup from '../shared/CustomPopup';
+import { theme } from '../../styles/theme';
+import Label from '../shared/Label';
 
 const PopupTxView = ({ isAccountPopup, popupStyle, offset }) => {
   const { gameEditionView } = useContext(GameEditionContext);
