+++ src/components/modals/swap-modals/SwapSuccesTxView.js
@@ -5,11 +5,11 @@ import { extractDecimal, reduceBalance } from '../../../utils/reduceBalance';
 import reduceToken from '../../../utils/reduceToken';
 import { getTokenIcon, showTicker } from '../../../utils/token-utils';
 import GameEditionLabel from '../../game-edition-v2/components/GameEditionLabel';
-import { Row, SuccessViewContainerGE, SuccesViewContainer } from './common-result-components';
+import { Row, SuccessViewContainerGE, SuccesViewContainer } from '../common-result-components';
 import { ChainIcon } from '../../../assets';
 import { ENABLE_GAS_STATION, GAS_PRICE } from '../../../constants/contextConstants';
 import { Divider, Icon } from 'semantic-ui-react';
-import PopupTxView from './PopupTxView';
+import PopupTxView from '../PopupTxView';
 import { theme } from '../../../styles/theme';
 import { GameEditionContext } from '../../../contexts/GameEditionContext';
 import { ApplicationContext } from '../../../contexts/ApplicationContext';
