+++ src/containers/PoolContainer.js
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import styled, { css } from 'styled-components/macro';
-import LiquidityContainer from './liquidity/LiquidityContainer';
-import LiquidityList from './liquidity/LiquidityList';
-import RemoveLiqContainer from './liquidity/RemoveLiqContainer';
+import LiquidityContainer from '../components/liquidity/LiquidityContainer';
+import LiquidityList from '../components/liquidity/LiquidityList';
+import RemoveLiqContainer from '../components/liquidity/RemoveLiqContainer';
 import modalBackground from '../assets/images/game-edition/modal-background.png';
 import { useGameEditionContext } from '../contexts';
 import useButtonScrollEvent from '../hooks/useButtonScrollEvent';
