+++ src/containers/SwapContainer.js
@@ -3,8 +3,8 @@ import React, { useContext, useEffect, useState } from 'react';
 import styled, { css } from 'styled-components/macro';
 import { throttle, debounce } from 'throttle-debounce';
 import { FadeIn } from '../components/shared/animations';
-import TxView from '../components/modals/swap-modals/TxView';
-import WalletRequestView from '../components/modals/swap-modals/WalletRequestView';
+import TxView from '../components/modals/TxView';
+import WalletRequestView from '../components/modals/WalletRequestView';
 import SwapButtonsForm from '../components/swap/SwapButtonsForm';
 import SwapForm from '../components/swap/SwapForm';
 import SwapResults from '../components/swap/SwapResults';
